refactor: import material data to handle shader rewrite (V1 and V2)
